vite and removing support code and adding comments
