+++ src/components/PokemonCollection.jsx
@@ -1,0 +1,12 @@
+// TODO: import and use the PokemonContext to retrieve the fetched pokemon
+function PokemonCollection(){
+    
+    // For each fetched pokemon, render a <PokemonCard />
+    return (
+        <div className="ui six cards">
+            {}
+        </div>
+    )
+}
+
+export default PokemonCollection
