--- src/context/PokemonContext.js
@@ -1,5 +1,0 @@
-import { createContext } from "react";
-
-const Context = createContext();
-
-export default Context;
