+++ src/context/PokemonContext.jsx
@@ -1,0 +1,1 @@
+// TODO: Create the context
