--- src/index.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './components/App';
-import PokemonProvider from './context/PokemonProvider'
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  <PokemonProvider>
-    <App />
-  </PokemonProvider> 
-);
