+++ src/main.jsx
@@ -1,0 +1,10 @@
+import ReactDOM from 'react-dom/client';
+import './index.css';
+import App from './App';
+
+// TODO: Import the PokemonProvider and wrap it around the App component
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+    <App />
+);
